Backport internal pull request 64
(API) Select experiment engine > select a client and one or more experiments > select a different client without clearing the client selection intermediately - API response error from Get Variables, that the experiment id is not found. This happens because the client id is first updated which causes the variables list to refresh (with the new client id and old experiment ids), before the experiment list is updated. This is temporarily fixed in the API by ignoring experiment ids that cannot be found when both client id and experiment ids are set, in the call to get experiment variables.